change int to size_t; #2004

+++ libweb3jsonrpc/Debug.cpp
@@ -122,7 +122,7 @@ Json::Value Debug::debug_getPatchTimestamps() {
     Json::Value jsonResponse;
     ChainParams chainParams = m_eth.chainParams();
 
-    int numberOfPatches = static_cast< int >( SchainPatchEnum::PatchesCount );
+    size_t numberOfPatches = static_cast< size_t >( SchainPatchEnum::PatchesCount );
     for ( size_t patch = 0; patch < numberOfPatches; patch++ ) {
         SchainPatchEnum patchEnum = static_cast< SchainPatchEnum >( patch );
         std::string patchName = getPatchNameForEnum( patchEnum ) + "Timestamp";
